changes in Invoice int to long.

+++ AddOptimization.Contracts/Dto/InvoiceDetailDto.cs
@@ -6,7 +6,7 @@ namespace AddOptimization.Contracts.Dto
     public class InvoiceDetailDto
     {
         public Guid Id { get; set; }
-        public int InvoiceId { get; set; }
+        public long InvoiceId { get; set; }
         public string Description { get; set; }
         public decimal Quantity { get; set; }
         public decimal UnitPrice { get; set; }
